Time: 5 ms (16.83%), Space: 6 MB (44.18%) - LeetHub

+++ 50-powx-n/50-powx-n.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     double myPow(double x, int n) {
-        int ans;
+        double ans;
         if(n==0){
             return 1;
         }
